requirements.txt
fastapi
pika
psycopg2-binary
redis
websockets
uvicorn
arcgis